+++ CHANGELOG.md
@@ -6,7 +6,23 @@
 
     *Nishiki Liu*
 
+## 2.26.1
+
+* Fix bug that raises when trying to use a collection before the component has been compiled.
+
+    *Blake Williams*
+
+## 2.26.0
+
+* Lazily evaluate component `content` in `render?`, preventing the `content` block from being evaluated when `render?` returns false.
+
+    *Blake Williams*
+
 * Do not generate template when using `--inline` flag.
+
+    *Hans Lemuet*
+
+* Add `--inline` option to the Haml and Slim generators
 
     *Hans Lemuet*